feat(styling): starts to change styling of app
