+++ src/index.css
@@ -1,3 +1,7 @@
 body, html, h1, h2, h3, h4, h5, h6, p, a {
   font-family: 'IBM Plex Sans', sans-serif
 }
+
+.Site-content, #root {
+  height: 100%
+}
